Refactored vector.reproject implicits to follow lib design pattern.
