+++ raster-test/src/test/scala/geotrellis/raster/reproject/RowTransformCheck.scala
@@ -1,10 +1,9 @@
 package geotrellis.raster.reproject
 
-import geotrellis.raster._
-import geotrellis.vector._
-import geotrellis.vector.reproject._
-import geotrellis.raster.testkit._
 import geotrellis.proj4._
+import geotrellis.raster._
+import geotrellis.raster.testkit._
+import geotrellis.vector._
 
 import org.scalacheck._
 import Prop._
@@ -16,7 +15,7 @@ import spire.syntax.cfor._
 trait RowTransformCheck { self: Properties =>
   case class TestCase(extent: Extent, srcX: Array[Double], srcY: Array[Double])
 
-  def genPoint(xmin: Double, ymin: Double, 
+  def genPoint(xmin: Double, ymin: Double,
                xmax: Double, ymax: Double): Gen[Point] =
     for {
       x <- choose(xmin, xmax)
@@ -25,7 +24,7 @@ trait RowTransformCheck { self: Properties =>
 }
 
 object RowTransformCheck_LatLngToWebMercator extends Properties("RowTransform") with RowTransformCheck {
-  lazy val genExtent: Gen[Extent] = 
+  lazy val genExtent: Gen[Extent] =
     for {
       p1 <- genPoint(-180, -85.0, 180.00, 85.0)
       p2 <- genPoint(-180, -85.0, 180.00, 85.0)
@@ -33,7 +32,7 @@ object RowTransformCheck_LatLngToWebMercator extends Properties("RowTransform") 
         val (x1, y1) = (p1.x, p1.y)
       val (x2, y2) = (p2.x, p2.y)
 
-      val (xmin, xmax) = 
+      val (xmin, xmax) =
         if(x1 < x2) (x1, x2) else (x2, x1)
 
       val (ymin, ymax) =
@@ -57,7 +56,7 @@ object RowTransformCheck_LatLngToWebMercator extends Properties("RowTransform") 
     Arbitrary(genTestCase)
 
   case class Threshold(v: Double)
-  lazy val genThreshold: Gen[Threshold] = 
+  lazy val genThreshold: Gen[Threshold] =
     for {
       v <- choose(0.1, 5.0)
     } yield Threshold(v)
@@ -80,7 +79,7 @@ object RowTransformCheck_LatLngToWebMercator extends Properties("RowTransform") 
     val srcProjected = src.map(_.reproject(transform))
     val dest = destX.zip(destY).map { case (x, y) => Point(x, y) }
 
-    val result = 
+    val result =
       srcProjected.zip(dest)
         .map { case(p1, p2) =>
           val dx = math.abs(p1.x - p2.x)
@@ -102,7 +101,7 @@ object RowTransformCheck_LatLngToWebMercator extends Properties("RowTransform") 
 
 object RowTransformCheck_UTMToWebMercator extends Properties("RowTransform") with RowTransformCheck {
   //Bounds: -78.0000, 0.0000, -72.0000, 84.0000
-  lazy val genExtent: Gen[Extent] = 
+  lazy val genExtent: Gen[Extent] =
     for {
       p1 <- genPoint(-77.9,0.0,-72.1,83.9)
       p2 <- genPoint(-77.9,0.0,-72.1,83.9)
@@ -110,7 +109,7 @@ object RowTransformCheck_UTMToWebMercator extends Properties("RowTransform") wit
       val (x1, y1) = (p1.x, p1.y)
       val (x2, y2) = (p2.x, p2.y)
 
-      val (xmin, xmax) = 
+      val (xmin, xmax) =
         if(x1 < x2) (x1, x2) else (x2, x1)
 
       val (ymin, ymax) =
@@ -134,7 +133,7 @@ object RowTransformCheck_UTMToWebMercator extends Properties("RowTransform") wit
     Arbitrary(genTestCase)
 
   case class Threshold(v: Double)
-  lazy val genThreshold: Gen[Threshold] = 
+  lazy val genThreshold: Gen[Threshold] =
     for {
       v <- choose(0.0, 2.0)
     } yield Threshold(v)
