+++ raster-test/src/test/scala/geotrellis/raster/reproject/RowTransformSpec.scala
@@ -2,7 +2,6 @@ package geotrellis.raster.reproject
 
 import geotrellis.raster._
 import geotrellis.vector._
-import geotrellis.vector.reproject._
 import geotrellis.vector.io.json._
 import geotrellis.raster.testkit._
 import geotrellis.proj4._
@@ -64,7 +63,7 @@ class RowTransformSpec extends FunSpec
     }
 
     it("should approximate WebMercator -> LatLng for raster extent") {
-      val Raster(expected, expectedExtent) = 
+      val Raster(expected, expectedExtent) =
         SinglebandGeoTiff("raster-test/data/reproject/slope_wsg84-nearestneighbor.tif").raster
 
       val re = RasterExtent(expected, expectedExtent)
@@ -94,7 +93,7 @@ class RowTransformSpec extends FunSpec
     }
 
     it("should approximate EPSG:32618 -> WebMercator for raster extent") {
-      val Raster(expected, expectedExtent) = 
+      val Raster(expected, expectedExtent) =
         SinglebandGeoTiff("raster-test/data/reproject/slope_epsg32618.tif").raster
 
       val re = RasterExtent(expected, expectedExtent)
