+++ spark/src/main/scala/geotrellis/spark/tiling/package.scala
@@ -1,7 +1,6 @@
 package geotrellis.spark
 
-import geotrellis.vector.Extent
-import geotrellis.vector.reproject._
+import geotrellis.vector._
 import geotrellis.proj4._
 
 /**
