+++ vector-test/src/test/scala/spec/geotrellis/vector/reproject/packageSpec.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2014 Azavea.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,9 +16,8 @@
 
 package geotrellis.vector
 
-import geotrellis.vector.reproject._
-import geotrellis.vector.io.json._
 import geotrellis.proj4._
+import geotrellis.vector.io.json._
 import geotrellis.vector.testkit._
 
 import org.scalatest._
